requirements.txt
pandas
python-dotenv
requests
tabulate